graph: refactor initial dependencies to reduce complexity
There was significant waste in breaking out weighting to different
concrete implementations of edges. The distrinction was removed and
weight is now a core property of an edge.

To reduce boiler plate, edges were renamed to lines and arrows
(undirected, directed). These may be constructed directly, with or
without weights. If constructed without weight, a default of `true` is
used for the weight value, meaning "yes, this link exists."